+++ packages/plugin-proxy/test/e2e/fixtures/host-2/db/users.js
@@ -1,0 +1,14 @@
+const USERS = [
+  {
+    id: 1,
+    name: "John Doe 2",
+  },
+  {
+    id: 2,
+    name: "Jane Doe 2",
+  },
+];
+
+module.exports = {
+  USERS,
+};