+++ packages/plugin-proxy/test/e2e/fixtures/host/mocks.js
@@ -1,0 +1,6 @@
+module.exports = [
+  {
+    id: "base",
+    routesVariants: ["get-users:success", "get-user:real"],
+  },
+];